requirements.txt
langchain~=0.0.275
PyPDF2~=3.0.1
python-dotenv~=1.0.0
streamlit~=1.26.0

torch~=2.0.1
transformers~=4.32.1
datasets~=2.14.4
tqdm~=4.66.1
numpy~=1.25.2
setuptools~=65.5.1
packaging~=23.1
accelerate~=0.22.0
safetensors~=0.3.3
pandas~=2.0.3